deskulpt_settings::model

Struct WidgetSettings

pub struct WidgetSettings {
    pub x: i32,
    pub y: i32,
    pub width: u32,
    pub height: u32,
    pub opacity: u8,
    pub z_index: i16,
    pub is_loaded: bool,
}

Per-widget settings.

Fields
x: i32

The leftmost x-coordinate in pixels.

y: i32

The topmost y-coordinate in pixels.

width: u32

The width in pixels.

height: u32

The height in pixels.

opacity: u8

The opacity in percentage.

z_index: i16

The z-index.

Higher z-index means the widget will be rendered above those with lower z-index. Widgets with the same z-index can have arbitrary rendering order. The allowed range is from -999 to 999.

is_loaded: bool

Whether the widget should be loaded on the canvas or not.

Implementations

impl WidgetSettings

fn deserialize_opacity<'de, D>(deserializer: D) -> Result<u8, D::Error>
where D: Deserializer<'de>,

Deserialization helper for opacity.

On error deserializing this field, it will be set to default (100). The deserialized value will be clamped to [1, 100].

impl WidgetSettings

pub fn apply_patch(&mut self, patch: WidgetSettingsPatch) -> bool

Apply a WidgetSettingsPatch.

This method also returns whether the widget settings is actually changed by the patch.